requirements_video.txt
# Dependências para o Módulo de Vídeo - AI Shorts
# Fase 1 - Setup técnico do Módulo 8

# YouTube content extraction
yt-dlp>=2024.1.1

# Video editing and processing
moviepy>=1.0.3
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# NLP processing
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.0.0

# CLIP model and machine learning
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Audio processing
pydub>=0.25.0
librosa>=0.10.0

# Additional utilities
requests>=2.31.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
tqdm>=4.66.0
python-dotenv>=1.0.0

# Optional: For better performance
numba>=0.58.0